Refactor to prepare for Store migration

+++ internal/overlord/state/task.go
@@ -29,8 +29,8 @@ import (
 
 type progress struct {
 	Label string `json:"label"`
-	Done  int    `json:"done"`
-	Total int    `json:"total"`
+	Done  int64  `json:"done"`
+	Total int64  `json:"total"`
 }
 
 // Task represents an individual operation to be performed
@@ -336,7 +336,7 @@ func (t *Task) Change() *Change {
 // Progress returns the current progress for the task.
 // If progress is not explicitly set, it returns
 // (0, 1) if the status is DoStatus and (1, 1) otherwise.
-func (t *Task) Progress() (label string, done, total int) {
+func (t *Task) Progress() (label string, done, total int64) {
 	t.state.reading()
 	if t.progress == nil {
 		if t.Status() == DoStatus {
@@ -348,7 +348,7 @@ func (t *Task) Progress() (label string, done, total int) {
 }
 
 // SetProgress sets the task progress to cur out of total steps.
-func (t *Task) SetProgress(label string, done, total int) {
+func (t *Task) SetProgress(label string, done, total int64) {
 	// Only mark state for checkpointing if progress is final.
 	if total > 0 && done == total {
 		t.state.writing()
