+++ internal/progress/reporter.go
@@ -2,5 +2,5 @@ package progress
 
 type Reporter struct {
 	Name   string
-	Report func(label string, done, total int)
+	Report func(label string, done, total int64)
 }
